Refactor: ActivityTopのディレクトリを変更
