+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/TopActivity.kt
@@ -1,9 +1,10 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.ui
 
 import androidx.appcompat.app.AppCompatActivity
+import jp.co.yumemi.android.code_check.R
 import java.util.*
 
 /**
