+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/detail/RepositoryDetailFragment.kt
@@ -10,7 +10,7 @@ import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.navArgs
 import coil.load
 import jp.co.yumemi.android.code_check.R
-import jp.co.yumemi.android.code_check.TopActivity.Companion.lastSearchDate
+import jp.co.yumemi.android.code_check.ui.TopActivity.Companion.lastSearchDate
 import jp.co.yumemi.android.code_check.databinding.RepositoryDetailFragmentBinding
 
 /**
