+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/search/SearchRepositoryViewModel.kt
@@ -10,7 +10,7 @@ import io.ktor.client.engine.android.*
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import java.util.*
-import jp.co.yumemi.android.code_check.TopActivity.Companion.lastSearchDate
+import jp.co.yumemi.android.code_check.ui.TopActivity.Companion.lastSearchDate
 import jp.co.yumemi.android.code_check.model.RepositoryInfo
 import kotlinx.coroutines.*
 import org.json.JSONObject
