+++ app/src/main/res/layout/search_repository_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TopActivity">
+    tools:context=".ui.TopActivity">
 
     <com.google.android.material.card.MaterialCardView
         android:id="@+id/searchBar"
